% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_codons_to_aa.R
\name{dna_codons_to_aa}
\alias{dna_codons_to_aa}
\title{translates dna codes to amino acids}
\usage{
dna_codons_to_aa(codons)
}
\arguments{
\item{codons}{a vector with strings containing dna codes e.g. "TTT"}
}
\value{
a sequence of amino acids
}
\description{
translates dna codes to amino acids
}
