% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_codons.R
\name{mk_codons}
\alias{mk_codons}
\title{R/mk_codons.R
Add together two numbers}
\usage{
mk_codons(dna, s = 1)
}
\arguments{
\item{codons}{is a substring of DNA}
}
\value{
Sequence of codons, length l
}
\description{
R/mk_codons.R
Add together two numbers
}
