% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_dna.R
\name{random_dna}
\alias{random_dna}
\title{Create random DNA sequence}
\usage{
random_dna(l)
}
\arguments{
\item{l}{Size of sequence}
}
\value{
Random sequence of DNA with the length \code{l}.
}
\description{
Create random DNA sequence
}
\examples{
random_dna(10)
random_dna(20)
}
