% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_rna.R
\name{random_rna}
\alias{random_rna}
\title{Generating random RNA sequence}
\usage{
random_rna(l)
}
\arguments{
\item{l}{size of sequence.}
}
\value{
random sequence of RNA.
}
\description{
Generating random RNA sequence
}
\examples{
random_rna(10)

}
